pyproject.toml
[tool.poetry]
name = "adat"
version = "0.2.0"
description = "datasets for fairness research"
authors = ["PhilippeRacicot <philippe.taharacicot@lacapitale.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.5.3"
google-cloud-storage = "^2.7.0"
scikit-learn = "^1.2.2"
feature-engine = "^1.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
